+++ docs/deploy.sh
@@ -308,11 +308,18 @@ redirects[ui-programming/react-native-styling-buttons]=tutorial/overview
 redirects[ui-programming/user-interface-libraries]=tutorial/overview
 
 # After renaming "workflows" to "eas-workflows"
-redirects[workflows/get-started]=eas-workflows/get-started
-redirects[workflows/triggers]=eas-workflows/triggers
-redirects[workflows/jobs]=eas-workflows/jobs
-redirects[workflows/control-flow]=eas-workflows/control-flow
-redirects[workflows/variables]=eas-workflows/variables
+redirects[workflows/get-started]=eas/workflows/get-started
+redirects[workflows/triggers]=eas/workflows/syntax/#on
+redirects[workflows/jobs]=eas/workflows/syntax/#jobs
+redirects[workflows/control-flow]=eas/workflows/syntax/#control-flow
+redirects[workflows/variables]=eas/workflows/syntax/#jobsjob_idoutputs
+
+# After moving eas-workflows to eas/workflows
+redirects[eas-workflows/get-started]=eas/workflows/get-started
+redirects[eas-workflows/triggers]=eas/workflows/syntax/#on
+redirects[eas-workflows/jobs]=eas/workflows/syntax/#jobs
+redirects[eas-workflows/control-flow]=eas/workflows/syntax/#control-flow
+redirects[eas-workflows/variables]=eas/workflows/syntax/#jobsjob_idoutputs
 
 # After adding distribution section under EAS
 redirects[distribution/publishing-websites]=guides/publishing-websites
@@ -330,6 +337,9 @@ redirects[bare]=bare/overview
 redirects[accounts/working-together]=accounts/account-types
 redirects[versions/latest/sdk/random]=versions/latest/sdk/crypto
 redirects[eas-update/known-issues]=eas-update/introduction
+
+# After consolidating the "Internal distribution" information
+redirects[guides/sharing-preview-releases]=build/internal-distribution
 
 echo "::group::[5/5] Add custom redirects"
 for i in "${!redirects[@]}" # iterate over keys