+++ packages/expo-modules-core/ios/Core/Promise.swift
@@ -4,6 +4,7 @@ public struct Promise: AnyArgument {
   public typealias ResolveClosure = (Any?) -> Void
   public typealias RejectClosure = (Exception) -> Void
 
+  internal weak var appContext: AppContext?
   public var resolver: ResolveClosure
   public var rejecter: RejectClosure
 
@@ -18,6 +19,15 @@ public struct Promise: AnyArgument {
   }
 
   public func resolve(_ value: Any? = nil) {
+    if let value = value as? AnySharedObject {
+      let result = Conversions.convertFunctionResult(
+        value,
+        appContext: appContext,
+        dynamicType: type(of: value).getDynamicType()
+      )
+      resolver(result)
+      return
+    }
     resolver(value)
   }