+++ packages/expo-ui/android/src/main/java/expo/modules/ui/PickerView.kt
@@ -1,7 +1,9 @@
 package expo.modules.ui
 
 import android.content.Context
+import android.graphics.Color
 import androidx.compose.material3.SegmentedButton
+import androidx.compose.material3.SegmentedButtonColors
 import androidx.compose.material3.SegmentedButtonDefaults
 import expo.modules.kotlin.viewevent.EventDispatcher
 import expo.modules.kotlin.views.ExpoComposeView
@@ -11,7 +13,6 @@ import androidx.compose.material3.RadioButton
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -25,11 +26,53 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.unit.dp
 import expo.modules.kotlin.AppContext
+import expo.modules.kotlin.records.Field
+import expo.modules.kotlin.records.Record
 import expo.modules.kotlin.views.ComposeProps
+import expo.modules.ui.button.ButtonColors
+
+class PickerColors : Record {
+  @Field
+  val activeBorderColor: Color? = null
+
+  @Field
+  val activeContentColor: Color? = null
+
+  @Field
+  val inactiveBorderColor: Color? = null
+
+  @Field
+  val inactiveContentColor: Color? = null
+
+  @Field
+  val disabledActiveBorderColor: Color? = null
+
+  @Field
+  val disabledActiveContentColor: Color? = null
+
+  @Field
+  val disabledInactiveBorderColor: Color? = null
+
+  @Field
+  val disabledInactiveContentColor: Color? = null
+
+  @Field
+  val activeContainerColor: Color? = null
+
+  @Field
+  val inactiveContainerColor: Color? = null
+
+  @Field
+  val disabledActiveContainerColor: Color? = null
+
+  @Field
+  val disabledInactiveContainerColor: Color? = null
+}
 
 data class PickerProps(
   val options: MutableState<Array<String>> = mutableStateOf(emptyArray()),
   val selectedIndex: MutableState<Int?> = mutableStateOf(null),
+  val elementColors: MutableState<PickerColors> = mutableStateOf(PickerColors()),
   val variant: MutableState<String> = mutableStateOf("segmented")
 ) : ComposeProps
 
@@ -39,12 +82,14 @@ class PickerView(context: Context, appContext: AppContext) : ExpoComposeView<Pic
 
   init {
     setContent {
-      val selectedIndex by remember { props.selectedIndex }
-      val options by remember { props.options }
+      val (selectedIndex) = props.selectedIndex
+      val (options) = props.options
+      val (colors) = props.elementColors
       val (variant) = props.variant
 
       @Composable
       fun SegmentedComposable() {
+      DynamicTheme {
         SingleChoiceSegmentedButtonRow {
           options.forEachIndexed { index, label ->
             SegmentedButton(
@@ -56,46 +101,62 @@ class PickerView(context: Context, appContext: AppContext) : ExpoComposeView<Pic
                 onOptionSelected(mapOf("index" to index, "label" to label))
               },
               selected = index == selectedIndex,
-              label = { Text(label) }
+              label = { Text(label) },
+              colors = SegmentedButtonDefaults.colors(
+                activeBorderColor = colors.activeBorderColor.compose,
+                activeContentColor = colors.activeContentColor.compose,
+                inactiveBorderColor = colors.inactiveBorderColor.compose,
+                inactiveContentColor = colors.inactiveContentColor.compose,
+                disabledActiveBorderColor = colors.disabledActiveBorderColor.compose,
+                disabledActiveContentColor = colors.disabledActiveContentColor.compose,
+                disabledInactiveBorderColor = colors.disabledInactiveBorderColor.compose,
+                disabledInactiveContentColor = colors.disabledInactiveContentColor.compose,
+                activeContainerColor = colors.activeContainerColor.compose,
+                inactiveContainerColor = colors.inactiveContainerColor.compose,
+                disabledActiveContainerColor = colors.disabledActiveContainerColor.compose,
+                disabledInactiveContainerColor = colors.disabledInactiveContainerColor.compose
+              )
             )
           }
         }
       }
-
-      @Composable
-      fun RadioComposable() {
-        Column(Modifier.selectableGroup()) {
-          options.forEachIndexed { index, label ->
-            Row(
-              Modifier.fillMaxWidth()
-                .height(28.dp)
-                .selectable(
-                  selected = index == selectedIndex,
-                  onClick = {
-                    onOptionSelected(mapOf("index" to index, "label" to label))
-                  },
-                  role = Role.RadioButton
-                ),
-              verticalAlignment = Alignment.CenterVertically
-            ) {
-              RadioButton(
-                selected = index == selectedIndex,
-                onClick = null
-              )
-              Text(
-                text = label,
-                modifier = Modifier.padding(start = 12.dp)
-              )
-            }
           }
+        @Composable
+        fun RadioComposable() {
+            DynamicTheme {
+            Column(Modifier.selectableGroup()) {
+                options.forEachIndexed { index, label ->
+                    Row(
+                        Modifier.fillMaxWidth()
+                            .height(28.dp)
+                            .selectable(
+                                selected = index == selectedIndex,
+                                onClick = {
+                                    onOptionSelected(mapOf("index" to index, "label" to label))
+                                },
+                                role = Role.RadioButton
+                            ),
+                        verticalAlignment = Alignment.CenterVertically
+                    ) {
+                        RadioButton(
+                            selected = index == selectedIndex,
+                            onClick = null
+                        )
+                        Text(
+                            text = label,
+                            modifier = Modifier.padding(start = 12.dp)
+                        )
+                    }
+                }
+            }
         }
-      }
+            }
 
-      if (variant == "segmented") {
-        SegmentedComposable()
-      } else if (variant == "radio") {
-        RadioComposable()
-      }
+        if (variant == "segmented") {
+            SegmentedComposable()
+        } else if (variant == "radio") {
+            RadioComposable()
+        }
     }
   }
 }